add error for not using subclass
filename and contents variables are created in different ways depending on
whether the file was found and initialized or created anew. Adding an error
to the NoteFile parent class that will raise unless the instance was
created through a child class and therefore defined those variables their
own way

+++ spec/created_file_spec.rb
@@ -1,12 +1,12 @@
-require_relative '../lib/creatd_file.rb'
-require_relative './spec_helper.rb'
+# frozen_string_literal: true
+
+require_relative '../lib/created_file'
+require_relative './spec_helper'
 require 'fileutils'
 
 describe CreatedFile do
   before :each do
-    if Dir.exists?(TEST_DIR)
-      FileUtils.rm_rf(TEST_DIR)
-    end
+    FileUtils.rm_rf(TEST_DIR) if Dir.exist?(TEST_DIR)
     Dir.mkdir(TEST_DIR)
   end
 
@@ -14,30 +14,30 @@ describe CreatedFile do
     FileUtils.rm_rf(TEST_DIR)
   end
 
-  describe "self#initialize" do
-    it "should create a new file on disk" do
-      expect {
-        CreatedFile.new('test title',TEST_DIR)
-      }.to change {
+  describe 'self#initialize' do
+    it 'should create a new file on disk' do
+      expect do
+        CreatedFile.new('test title', TEST_DIR)
+      end.to change {
         files = Dir["#{TEST_DIR}/**/*"] { |file| File.file?(file) }
         files.length
       }.by 1
     end
 
-    it "should create two files even if in the same second" do
-      expect {
-        CreatedFile.new('test title',TEST_DIR)
-        CreatedFile.new('test title',TEST_DIR)
-      }.to change {
+    it 'should create two files even if in the same second' do
+      expect do
+        CreatedFile.new('test title', TEST_DIR)
+        CreatedFile.new('test title', TEST_DIR)
+      end.to change {
         files = Dir["#{TEST_DIR}/**/*"] { |file| File.file?(file) }
         files.length
       }.by 2
     end
 
-    it "should create the file with the given title" do
+    it 'should create the file with the given title' do
       CreatedFile.new('test title 123', TEST_DIR)
       text_files = Dir.new(TEST_DIR).entries.filter do |filename|
-        filename.match /\.md/
+        filename.match(/\.md/)
       end
 
       created_filename = text_files[0]
